Add rebuild clarification
I had Revit 2019 open when rebuilding the solution. The solution failed because it wasn't able to delete %APPDATA%\Autodesk\Revit\Addins\2019\BatchRvt

While this isn't a problem for the initial installation (since Revit won't be using the folder), this could be a problem in the future for those that want to upgrade their RBP version.